Agent key migration, DB fix, UI tabs/page cleanup
Add robust credential handling and migration for agent encryption keys: surface LoadError on startup, log clear guidance when creds fail to decrypt, try legacy (pre-1.6.14) machine key fallback and transparently re-encrypt under the new host-stable key. Fix machine-key derivation (make host-stable instead of user-stable) and encapsulate decryption/migration helpers.

Fix SQLite autoincrement issue: change ServerMetrics.id from BigInteger to Integer so SQLite will autoincrement rows and metrics are recorded.

Ensure background heartbeat checker can touch the DB by capturing the Flask app in AgentRegistry and entering an app_context in the heartbeat thread (with a test-friendly fallback for when no app is available).

Frontend UI: simplify Tabs component classes and update pages to use a unified .page-container wrapper; replace many manual tab button implementations with Tabs/TabsList/TabsTrigger and add improved tab styles (disabled/focus states). Misc: add docs/PLAN_AGENT_FLEET.md to .gitignore.

+++ backend/app/models/server.py
@@ -361,7 +361,13 @@ class ServerMetrics(db.Model):
     """Historical metrics from servers"""
     __tablename__ = 'server_metrics'
 
-    id = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
+    # SQLite only autoincrements when the column type is exactly INTEGER
+    # PRIMARY KEY (rowid alias). With BigInteger every insert failed with
+    # "NOT NULL constraint failed: server_metrics.id" and no metrics ever
+    # got stored — exactly the silent dashboard-zeroes symptom we just
+    # tracked down. Integer caps at ~2.1B rows which at 1 sample/sec is
+    # ~68 years per server, plenty.
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     server_id = db.Column(db.String(36), db.ForeignKey('servers.id'), nullable=False, index=True)
 
     timestamp = db.Column(db.DateTime, default=datetime.utcnow, index=True)
